feat: stream embeds via websocket
