+++ DemiCatPlugin/UiRenderer.cs
@@ -5,6 +5,10 @@ using System.Net.Http;
 using System.Text.Json;
 using System.Numerics;
 using System.Threading.Tasks;
+using System.IO;
+using System.Net.WebSockets;
+using System.Text;
+using System.Threading;
 using DiscordHelper;
 using Dalamud.Bindings.ImGui;
 
@@ -15,20 +19,181 @@ public class UiRenderer : IDisposable
     private readonly HttpClient _httpClient;
     private readonly Config _config;
     private readonly Dictionary<string, EventView> _embeds = new();
+    private readonly List<EmbedDto> _embedDtos = new();
     private string _channelId;
     private EventView? _current;
+    private ClientWebSocket? _webSocket;
+    private CancellationTokenSource? _pollCts;
 
     public UiRenderer(Config config, HttpClient httpClient)
     {
         _config = config;
         _httpClient = httpClient;
         _channelId = config.EventChannelId;
+
+        if (_config.Enabled)
+        {
+            StartPolling();
+            _ = ConnectWebSocket();
+        }
     }
 
     public string ChannelId
     {
         get => _channelId;
         set => _channelId = value;
+    }
+
+    private void StartPolling()
+    {
+        if (_pollCts != null)
+        {
+            return;
+        }
+        _pollCts = new CancellationTokenSource();
+        _ = PollLoop(_pollCts.Token);
+    }
+
+    private void StopPolling()
+    {
+        _pollCts?.Cancel();
+        _pollCts = null;
+    }
+
+    private async Task PollLoop(CancellationToken token)
+    {
+        var interval = TimeSpan.FromSeconds(_config.PollIntervalSeconds);
+        using var timer = new PeriodicTimer(interval);
+        try
+        {
+            while (await timer.WaitForNextTickAsync(token))
+            {
+                await PollEmbeds();
+                if (_webSocket == null || _webSocket.State != WebSocketState.Open)
+                {
+                    _ = ConnectWebSocket();
+                }
+            }
+        }
+        catch (OperationCanceledException)
+        {
+            // ignored
+        }
+    }
+
+    private async Task PollEmbeds()
+    {
+        try
+        {
+            var request = new HttpRequestMessage(HttpMethod.Get, $"{_config.ApiBaseUrl.TrimEnd('/')}/api/embeds");
+            if (!string.IsNullOrEmpty(_config.AuthToken))
+            {
+                request.Headers.Add("X-Api-Key", _config.AuthToken);
+            }
+            var response = await _httpClient.SendAsync(request);
+            if (!response.IsSuccessStatusCode)
+            {
+                return;
+            }
+            var stream = await response.Content.ReadAsStreamAsync();
+            var embeds = await JsonSerializer.DeserializeAsync<List<EmbedDto>>(stream) ?? new List<EmbedDto>();
+            _ = PluginServices.Framework.RunOnTick(() =>
+            {
+                _embedDtos.Clear();
+                _embedDtos.AddRange(embeds);
+                SetEmbeds(_embedDtos);
+            });
+        }
+        catch
+        {
+            // ignored
+        }
+    }
+
+    private async Task ConnectWebSocket()
+    {
+        if (_webSocket != null && _webSocket.State == WebSocketState.Open)
+        {
+            return;
+        }
+
+        try
+        {
+            _webSocket?.Dispose();
+            _webSocket = new ClientWebSocket();
+            if (!string.IsNullOrEmpty(_config.AuthToken))
+            {
+                _webSocket.Options.SetRequestHeader("X-Api-Key", _config.AuthToken);
+            }
+            var baseUrl = _config.ApiBaseUrl.TrimEnd('/');
+            var wsUrl = new Uri(($"{baseUrl}/ws/embeds")
+                .Replace("http://", "ws://")
+                .Replace("https://", "wss://"));
+            await _webSocket.ConnectAsync(wsUrl, CancellationToken.None);
+            StopPolling();
+            await ReceiveLoop();
+        }
+        catch
+        {
+            StartPolling();
+        }
+    }
+
+    private async Task ReceiveLoop()
+    {
+        if (_webSocket == null)
+        {
+            return;
+        }
+
+        var buffer = new byte[8192];
+        try
+        {
+            while (_webSocket.State == WebSocketState.Open)
+            {
+                var ms = new MemoryStream();
+                WebSocketReceiveResult result;
+                do
+                {
+                    result = await _webSocket.ReceiveAsync(buffer, CancellationToken.None);
+                    if (result.MessageType == WebSocketMessageType.Close)
+                    {
+                        await _webSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, string.Empty, CancellationToken.None);
+                        return;
+                    }
+                    ms.Write(buffer, 0, result.Count);
+                } while (!result.EndOfMessage);
+
+                var json = Encoding.UTF8.GetString(ms.ToArray());
+                var embed = JsonSerializer.Deserialize<EmbedDto>(json);
+                if (embed != null)
+                {
+                    _ = PluginServices.Framework.RunOnTick(() =>
+                    {
+                        var index = _embedDtos.FindIndex(e => e.Id == embed.Id);
+                        if (index >= 0)
+                        {
+                            _embedDtos[index] = embed;
+                        }
+                        else
+                        {
+                            _embedDtos.Add(embed);
+                        }
+                        SetEmbeds(_embedDtos);
+                    });
+                }
+            }
+        }
+        catch
+        {
+            // ignored
+        }
+        finally
+        {
+            _webSocket?.Dispose();
+            _webSocket = null;
+            StartPolling();
+        }
     }
 
     public void SetEmbeds(IEnumerable<EmbedDto> embeds)
@@ -70,7 +235,12 @@ public class UiRenderer : IDisposable
             }
             var stream = await response.Content.ReadAsStreamAsync();
             var embeds = await JsonSerializer.DeserializeAsync<List<EmbedDto>>(stream) ?? new List<EmbedDto>();
-            SetEmbeds(embeds);
+            _ = PluginServices.Framework.RunOnTick(() =>
+            {
+                _embedDtos.Clear();
+                _embedDtos.AddRange(embeds);
+                SetEmbeds(_embedDtos);
+            });
         }
         catch
         {
@@ -107,6 +277,24 @@ public class UiRenderer : IDisposable
 
     public void Dispose()
     {
+        StopPolling();
+        if (_webSocket != null)
+        {
+            try
+            {
+                if (_webSocket.State == WebSocketState.Open)
+                {
+                    _webSocket.CloseAsync(WebSocketCloseStatus.NormalClosure, string.Empty, CancellationToken.None).Wait();
+                }
+            }
+            catch
+            {
+                // ignored
+            }
+            _webSocket.Dispose();
+            _webSocket = null;
+        }
+
         foreach (var view in _embeds.Values)
         {
             view.Dispose();
